add tests, release notes, move text df to conftest
